checking some instances

+++ 16-binary_tree_is_perfect.c
@@ -108,5 +108,5 @@ int binary_tree_is_perfect(const binary_tree_t *tree)
     num_nodes = (1 << (left_height + 1)) - 1;
 
     /* Check if the number of nodes meets the perfect binary tree condition */
-    return (num_nodes == BT_size(tree));
+    return (num_nodes == (int)BT_size(tree));
 }
